+++ libs/naas-abi-marketplace/naas_abi_marketplace/applications/arxiv/agents/ArXivAgent_test.py
@@ -1,10 +1,17 @@
 import pytest
-from naas_abi_marketplace.applications.arxiv.agents.ArXivAgent import create_agent
 
 
 @pytest.fixture
 def agent():
-    return create_agent()
+    from naas_abi_core.engine.Engine import Engine
+    from naas_abi_marketplace.applications.arxiv.agents.ArXivAgent import (
+        ArXivAgent,
+    )
+
+    engine = Engine()
+    engine.load(module_names=["naas_abi_marketplace.applications.arxiv"])
+
+    return ArXivAgent.New()
 
 
 def test_agent_name(agent):